Refactor Container to delegate operations to an internal class.
A Container that exists will be backed by an ExistingContainer; one that does
not will be backed by a NullContainer. Containers are a little lazier about
things like fetching metadata.

This lets me implement delete() properly, along with always providing a
consistent interface, and not forcing clients to check exists() everywhere.